requirements.txt
#==== requirements.txt ====

sqlmodel>=0.0.16
telethon>=1.34
python-dotenv>=1.0
psycopg2-binary>=2.9.9
supabase>=2.0
py7zr>=0.20.0
httpx[http2]
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
psutil
tenacity>=9.1.2